requirements.txt
flask
flask-cors
Pillow
google-generativeai
matplotlib
numpy
sympy
python-dotenv
